requirements.txt
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
diff-match-patch==20121119
Django==2.0
django-admin-tools==0.8.1
django-admin-views==0.8.0
django-bootstrap3==9.1.0
django-import-export==0.6.1
django-multi-captcha-admin==1.0.0
django-phonenumber-field==1.3.0
django-recaptcha2==1.0.3
django-shortuuidfield==0.1.3
django-smalluuid==1.1.1
et-xmlfile==1.0.1

httplib2==0.10.3
idna==2.6
jdcal==1.3
jws==0.1.3
oauth2client==3.0.0
odfpy==1.3.6
olefile==0.44
openpyxl==2.4.9
phonenumberslite==8.8.8
Pillow==4.3.0
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2017.3
PyYAML==3.12
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
shortuuid==0.5.0
six==1.11.0
smalluuid==1.0.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
